+++ cmd/ingest-server/main.go
@@ -36,7 +36,7 @@ func main() {
 	s := server.NewServer(configManager)
 
 	mux := http.NewServeMux()
-	mux.Handle("POST /upload/{app}", s.RateLimitMiddleware(http.HandlerFunc(s.UploadHandler)))
+	mux.Handle("POST /upload/{app}", s.IPLimiter.RateLimitMiddleware(http.HandlerFunc(s.UploadHandler)))
 	mux.Handle("GET /version", http.HandlerFunc(s.VersionHandler))
 
 	srv := &http.Server{
